Add git setup function.
While I don't particularly like this generate on-the-fly approach and
would prefer to actually version the .gitconfig file, I have to many
different default user combinations to make that practical, so this will
have to suffice for the time being. It will simply ask you what it
should set as the global user.name and user.email values and then set
the aliases. Of course it is still possible to override any of these
values on a per repository basis.