Add printer admin function
This forwards the port 631 of the print server to port 3631 of the local
machine so that the cups administration page can be viewed in the local
browser. This simplifies the cups configuration since we don't need to
open any other ports or restrict any other access other than the
(already available) ssh port.